styled_Componets/Backend/requirements.txt
alembic==1.16.2
amqp==5.3.1
asgiref==3.8.1
audioop-lts==0.2.1
Authlib==1.5.2
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.11
cryptography==44.0.3
Django==5.0.1
django-extensions==4.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-nested-admin==4.0.2
djangorestframework==3.16.1
docker==7.1.0
dominate==2.9.1
Faker==37.3.0
Flask==3.1.0
Flask-Bootstrap4==4.0.2
flask-bootstrap5==0.1.dev1
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.17.0
mypy_extensions==1.1.0
packaging==24.1
pathspec==0.12.1
pillow==11.2.1
pip-autoremove==0.10.0
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.12.0
pyjokes==0.8.3
pypiwin32==223
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
redis==6.0.0
requests==2.32.3
sentry-sdk==2.27.0
setuptools==70.0.0
six==1.17.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlparse==0.4.2
standard-aifc==3.13.0
standard-chunk==3.13.0
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.2
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
